refactor(LottoGenerator): 로또 생성기 interface로 변경
