+++ README.md
@@ -94,8 +94,10 @@
     - [ ] 수동 로또 수 입력
         - [ ] 구입금액보다 수동으로 입력한 로또 금액이 큰 경우 오류
         - [ ] 숫자 관련 유효성 처리
+        - [ ] 구매 금액 입력 후 수동로또 구매 개수 입력 시 유효성 처리
     - [ ] 수동으로 구매할 번호 입력
         - [ ] 로또 수 만큼 순회하면서 입력 및 유효성 처리
         - [ ] 입력한 수동 로또 기억
+        - [ ] 수동로또, 자동로또 티켓에 담기 
     - [ ] 수동 구매 개수, 자동 구매 개수 출력
     - [ ] 사용자 입력 관련 예외 처리 시 에러가 발생하지 않도록 한다.
