+++ src/main/java/lotto/controller/LottoController.java
@@ -16,9 +16,14 @@ public class LottoController {
     }
 
     public void startLotto() {
-//        LottoMoney lottoMoney = InputView.getLottoPurchasePric1();
+
 
         int purchasePrice = InputView.getLottoPurchasePrice();
+//        String manualLottoCount = InputView.getManualLottoCount();
+        // TODO: 구매 금액 유효성 체크
+//        LottoMoney lottoMoney = InputView.getLottoPurchasePrice();
+        // TODO: 구매 금액 입력 후 수동로또 구매 시 구매 금액을 초과한 경우
+        // TODO: 수동 로또 번호 추가 로직
         LottoTickets lottoTickets = lottoGame.buy(purchasePrice);
 
         ResultView.lottoPurchase(lottoTickets.ticketCount(), lottoTickets.toString());
