+++ src/test/java/lotto/domain/LottoGeneratorTest.java
@@ -11,7 +11,8 @@ class LottoGeneratorTest {
     @DisplayName("로또 랜덤 숫자 사이즈 확인하기")
     @Test
     void check_lotto_random_number_size() {
-        LottoNumbers lottoNumbers = LottoGenerator.generateLottoNumber();
+        LottoGenerator lottoGenerator = new AutoLottoGenerator();
+        LottoNumbers lottoNumbers = lottoGenerator.generateLottoNumber();
         assertThat(lottoNumbers.numbers().size()).isEqualTo(6);
     }
